Fix sudden crash-proneness #4
Since October 2016 the ICS crashed within minutes, when games are going on.
The cause seems to be a reversed test for bughouse in game_ended(), which
informed players of the availability of the bughouse partners of the ending
game if and only if the game was not bughouse. Accessing bughouse partners
only when they exist seems to have solved the problem.