Hack to bypass timeseal decoder
The decription of timeseal data is now done by the ICS itself, rather
than by the external timeseal decoder process. The process is still started,
though, and the logic for testing whether it was initialized and should be
used is not altered. The ICS just refrains from ever sending any data
to it for decoding.